// src/generated/livekit_egress.proto.ts
  export type UNRECOGNIZED = typeof EncodedFileType.UNRECOGNIZED;
// ...
export const EncodingOptionsPreset = {
  /** H264_720P_30 - 1280x720, 30fps, 3000kpbs, H.264_MAIN / OPUS */
  H264_720P_30: 0,
  /** H264_720P_60 - 1280x720, 60fps, 4500kbps, H.264_MAIN / OPUS */
  H264_720P_60: 1,
  /** H264_1080P_30 - 1920x1080, 30fps, 4500kbps, H.264_MAIN / OPUS */
  H264_1080P_30: 2,
  /** H264_1080P_60 - 1920x1080, 60fps, 6000kbps, H.264_MAIN / OPUS */
  H264_1080P_60: 3,
  /** PORTRAIT_H264_720P_30 - 720x1280, 30fps, 3000kpbs, H.264_MAIN / OPUS */
  PORTRAIT_H264_720P_30: 4,
  /** PORTRAIT_H264_720P_60 - 720x1280, 60fps, 4500kbps, H.264_MAIN / OPUS */
  PORTRAIT_H264_720P_60: 5,
  /** PORTRAIT_H264_1080P_30 - 1080x1920, 30fps, 4500kbps, H.264_MAIN / OPUS */
  PORTRAIT_H264_1080P_30: 6,
  /** PORTRAIT_H264_1080P_60 - 1080x1920, 60fps, 6000kbps, H.264_MAIN / OPUS */
  PORTRAIT_H264_1080P_60: 7,
  UNRECOGNIZED: -1,
} as const;

export type EncodingOptionsPreset = typeof EncodingOptionsPreset[keyof typeof EncodingOptionsPreset];
// ...
  export type H264_720P_30 = typeof EncodingOptionsPreset.H264_720P_30;
  export type H264_720P_60 = typeof EncodingOptionsPreset.H264_720P_60;
  export type H264_1080P_30 = typeof EncodingOptionsPreset.H264_1080P_30;
  export type H264_1080P_60 = typeof EncodingOptionsPreset.H264_1080P_60;
  export type PORTRAIT_H264_720P_30 = typeof EncodingOptionsPreset.PORTRAIT_H264_720P_30;
  export type PORTRAIT_H264_720P_60 = typeof EncodingOptionsPreset.PORTRAIT_H264_720P_60;
  export type PORTRAIT_H264_1080P_30 = typeof EncodingOptionsPreset.PORTRAIT_H264_1080P_30;
  export type PORTRAIT_H264_1080P_60 = typeof EncodingOptionsPreset.PORTRAIT_H264_1080P_60;
// ...
export const EgressStatus = {
  EGRESS_STARTING: 0,
  EGRESS_ACTIVE: 1,
  EGRESS_ENDING: 2,
  EGRESS_COMPLETE: 3,
  EGRESS_FAILED: 4,
  EGRESS_ABORTED: 5,
  EGRESS_LIMIT_REACHED: 6,
  UNRECOGNIZED: -1,
} as const;

export type EgressStatus = typeof EgressStatus[keyof typeof EgressStatus];
// ...
  export type EGRESS_STARTING = typeof EgressStatus.EGRESS_STARTING;
  export type EGRESS_ACTIVE = typeof EgressStatus.EGRESS_ACTIVE;
  export type EGRESS_ENDING = typeof EgressStatus.EGRESS_ENDING;
// ...
export const EgressSourceType = { EGRESS_SOURCE_TYPE_WEB: 0, EGRESS_SOURCE_TYPE_SDK: 1, UNRECOGNIZED: -1 } as const;

export type EgressSourceType = typeof EgressSourceType[keyof typeof EgressSourceType];
// ...
  export type EGRESS_SOURCE_TYPE_WEB = typeof EgressSourceType.EGRESS_SOURCE_TYPE_WEB;
  export type EGRESS_SOURCE_TYPE_SDK = typeof EgressSourceType.EGRESS_SOURCE_TYPE_SDK;
// ...
export interface RoomCompositeEgressRequest {
  /** required */
  roomName: string;
  /** (optional) */
  layout: string;
  /** (default false) */
  audioOnly: boolean;
  /** only applies to audio_only egress (default DEFAULT_MIXING) */
  audioMixing: AudioMixing;
  /** (default false) */
  videoOnly: boolean;
  /** template base url (default https://recorder.livekit.io) */
  customBaseUrl: string;
  /** @deprecated */
  file?:
    | EncodedFileOutput
    | undefined;
  /** @deprecated */
  stream?:
    | StreamOutput
    | undefined;
  /** @deprecated */
  segments?:
    | SegmentedFileOutput
    | undefined;
  /** (default H264_720P_30) */
  preset?:
    | EncodingOptionsPreset
    | undefined;
  /** (optional) */
  advanced?: EncodingOptions | undefined;
  fileOutputs: EncodedFileOutput[];
  streamOutputs: StreamOutput[];
  segmentOutputs: SegmentedFileOutput[];
  imageOutputs: ImageOutput[];
}

/** record any website */
export interface WebEgressRequest {
  url: string;
  audioOnly: boolean;
  videoOnly: boolean;
  awaitStartSignal: boolean;
  /** @deprecated */
  file?:
    | EncodedFileOutput
    | undefined;
  /** @deprecated */
  stream?:
    | StreamOutput
    | undefined;
  /** @deprecated */
  segments?: SegmentedFileOutput | undefined;
  preset?: EncodingOptionsPreset | undefined;
  advanced?: EncodingOptions | undefined;
  fileOutputs: EncodedFileOutput[];
  streamOutputs: StreamOutput[];
  segmentOutputs: SegmentedFileOutput[];
  imageOutputs: ImageOutput[];
}

/** record audio and video from a single participant */
export interface ParticipantEgressRequest {
  /** required */
  roomName: string;
  /** required */
  identity: string;
  /** (default false) */
  screenShare: boolean;
  /** (default H264_720P_30) */
  preset?:
    | EncodingOptionsPreset
    | undefined;
  /** (optional) */
  advanced?: EncodingOptions | undefined;
  fileOutputs: EncodedFileOutput[];
  streamOutputs: StreamOutput[];
  segmentOutputs: SegmentedFileOutput[];
  imageOutputs: ImageOutput[];
}

/** containerize up to one audio and one video track */
export interface TrackCompositeEgressRequest {
  /** required */
  roomName: string;
  /** (optional) */
  audioTrackId: string;
  /** (optional) */
  videoTrackId: string;
  /** @deprecated */
  file?:
    | EncodedFileOutput
    | undefined;
  /** @deprecated */
  stream?:
    | StreamOutput
    | undefined;
  /** @deprecated */
  segments?:
    | SegmentedFileOutput
    | undefined;
  /** (default H264_720P_30) */
  preset?:
    | EncodingOptionsPreset
    | undefined;
  /** (optional) */
  advanced?: EncodingOptions | undefined;
  fileOutputs: EncodedFileOutput[];
  streamOutputs: StreamOutput[];
  segmentOutputs: SegmentedFileOutput[];
  imageOutputs: ImageOutput[];
}

/** record tracks individually, without transcoding */
export interface TrackEgressRequest {
  /** required */
  roomName: string;
  /** required */
  trackId: string;
  file?: DirectFileOutput | undefined;
  websocketUrl?: string | undefined;
}
// ...
export interface EgressInfo {
  egressId: string;
  roomId: string;
  roomName: string;
  sourceType: EgressSourceType;
  status: EgressStatus;
  startedAt: number;
  endedAt: number;
  updatedAt: number;
  details: string;
  error: string;
  errorCode: number;
  roomComposite?: RoomCompositeEgressRequest | undefined;
  web?: WebEgressRequest | undefined;
  participant?: ParticipantEgressRequest | undefined;
  trackComposite?: TrackCompositeEgressRequest | undefined;
  track?:
    | TrackEgressRequest
    | undefined;
  /** @deprecated */
  stream?:
    | StreamInfoList
    | undefined;
  /** @deprecated */
  file?:
    | FileInfo
    | undefined;
  /** @deprecated */
  segments?: SegmentsInfo | undefined;
  streamResults: StreamInfo[];
  fileResults: FileInfo[];
  segmentResults: SegmentsInfo[];
  imageResults: ImagesInfo[];
  manifestLocation: string;
  /** next ID: 27 */
  backupStorageUsed: boolean;
}

/** @deprecated */
export interface StreamInfoList {
  info: StreamInfo[];
}

export interface StreamInfo {
  url: string;
  startedAt: number;
  endedAt: number;
  duration: number;
  status: StreamInfo_Status;
  error: string;
}
// ...
export interface FileInfo {
  filename: string;
  startedAt: number;
  endedAt: number;
  duration: number;
  size: number;
  location: string;
}

export interface SegmentsInfo {
  playlistName: string;
  livePlaylistName: string;
  duration: number;
  size: number;
  playlistLocation: string;
  livePlaylistLocation: string;
  segmentCount: number;
  startedAt: number;
  endedAt: number;
}

export interface ImagesInfo {
  filenamePrefix: string;
  imageCount: number;
  startedAt: number;
  endedAt: number;
}
// ...
export const TrackEgressRequest: MessageFns<TrackEgressRequest> = {
  encode(message: TrackEgressRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.roomName !== "") {
      writer.uint32(10).string(message.roomName);
    }
    if (message.trackId !== "") {
      writer.uint32(18).string(message.trackId);
    }
    if (message.file !== undefined) {
      DirectFileOutput.encode(message.file, writer.uint32(26).fork()).join();
    }
    if (message.websocketUrl !== undefined) {
      writer.uint32(34).string(message.websocketUrl);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): TrackEgressRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseTrackEgressRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.roomName = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.trackId = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.file = DirectFileOutput.decode(reader, reader.uint32());
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.websocketUrl = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): TrackEgressRequest {
    return {
      roomName: isSet(object.roomName) ? globalThis.String(object.roomName) : "",
      trackId: isSet(object.trackId) ? globalThis.String(object.trackId) : "",
      file: isSet(object.file) ? DirectFileOutput.fromJSON(object.file) : undefined,
      websocketUrl: isSet(object.websocketUrl) ? globalThis.String(object.websocketUrl) : undefined,
    };
  },

  toJSON(message: TrackEgressRequest): unknown {
    const obj: any = {};
    if (message.roomName !== "") {
      obj.roomName = message.roomName;
    }
    if (message.trackId !== "") {
      obj.trackId = message.trackId;
    }
    if (message.file !== undefined) {
      obj.file = DirectFileOutput.toJSON(message.file);
    }
    if (message.websocketUrl !== undefined) {
      obj.websocketUrl = message.websocketUrl;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<TrackEgressRequest>, I>>(base?: I): TrackEgressRequest {
    return TrackEgressRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<TrackEgressRequest>, I>>(object: I): TrackEgressRequest {
    const message = createBaseTrackEgressRequest();
    message.roomName = object.roomName ?? "";
    message.trackId = object.trackId ?? "";
    message.file = (object.file !== undefined && object.file !== null)
      ? DirectFileOutput.fromPartial(object.file)
      : undefined;
    message.websocketUrl = object.websocketUrl ?? undefined;
    return message;
  },
};
// ...
export const StreamInfoList: MessageFns<StreamInfoList> = {
  encode(message: StreamInfoList, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    for (const v of message.info) {
      StreamInfo.encode(v!, writer.uint32(10).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): StreamInfoList {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseStreamInfoList();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.info.push(StreamInfo.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): StreamInfoList {
    return { info: globalThis.Array.isArray(object?.info) ? object.info.map((e: any) => StreamInfo.fromJSON(e)) : [] };
  },

  toJSON(message: StreamInfoList): unknown {
    const obj: any = {};
    if (message.info?.length) {
      obj.info = message.info.map((e) => StreamInfo.toJSON(e));
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<StreamInfoList>, I>>(base?: I): StreamInfoList {
    return StreamInfoList.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<StreamInfoList>, I>>(object: I): StreamInfoList {
    const message = createBaseStreamInfoList();
    message.info = object.info?.map((e) => StreamInfo.fromPartial(e)) || [];
    return message;
  },
};